change all remaining tables
