+++ src/components/dashboard/containers/AddItem/AddItem.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import AppBar from '@material-ui/core/AppBar'
 import Tabs from '@material-ui/core/Tabs'
 import Tab from '@material-ui/core/Tab'
-import EnhancedTable from 'components/dashboard/containers/Tables/EnhancedTablele'
+import { AdUnitsTable } from 'components/dashboard/containers/Tables'
 import Translate from 'components/translate/Translate'
 import { ContentBox, ContentBody } from 'components/common/dialog/content'
 import { withStyles } from '@material-ui/core/styles'
@@ -53,14 +53,12 @@ export class AddItem extends Component {
 
 						{tabIndex === 1 && (
 							<ContentBody>
-								<EnhancedTable
-									{...rest}
-									objModel={this.props.objModel}
-									parentItem={this.props.addTo}
-									addToItem
+								<AdUnitsTable
 									items={this.props.items}
-									viewModeId={this.props.viewMode}
-									listMode={this.props.listMode}
+									// noSearch
+									noActions
+									noDownload
+									noPrint
 								/>
 							</ContentBody>
 						)}
