+++ src/components/dashboard/containers/Tables/AdUnitsTable.js
@@ -6,7 +6,7 @@ import { Visibility } from '@material-ui/icons'
 import Img from 'components/common/img/Img'
 import MUIDataTableEnchanced from 'components/dashboard/containers/Tables/MUIDataTableEnchanced'
 import { withReactRouterLink } from 'components/common/rr_hoc/RRHoc'
-import { t, selectAdUnits, selectSide } from 'selectors'
+import { t, selectAdUnits, selectSide, selectNewCampaign } from 'selectors'
 import { makeStyles } from '@material-ui/core/styles'
 import { useSelector } from 'react-redux'
 import { styles } from './styles'
@@ -21,7 +21,18 @@ function AdUnitsTable(props) {
 	const classes = useStyles()
 	const adUnits = useSelector(selectAdUnits)
 	const side = useSelector(selectSide)
-	const data = Object.values(props.items || adUnits).map(item => ({
+	const [selected, setSelected] = React.useState([])
+	const {
+		items, //
+		noActions,
+		noClone,
+		campaignUnits,
+		validate,
+		handleSelect,
+	} = props
+
+	const data = Object.values(items || adUnits).map(item => ({
+		id: item.id,
 		media: {
 			id: item.id,
 			mediaUrl: item.mediaUrl,
@@ -35,6 +46,22 @@ function AdUnitsTable(props) {
 			id: item.id || item.ipfs,
 		},
 	}))
+
+	React.useEffect(() => {
+		const selectedItems = selected.map(i => data[i].id)
+		const isValid = !!selectedItems.length
+		if (validate && handleSelect) {
+			validate('adUnits', {
+				isValid: isValid,
+				err: { msg: 'ERR_ADUNITS_REQUIRED' },
+				dirty: true,
+			})
+			isValid && handleSelect(selectedItems)
+		}
+		// TODO: temp fix - need to redesign the table again
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [selected])
+
 	const columns = [
 		{
 			name: 'media',
@@ -87,7 +114,7 @@ function AdUnitsTable(props) {
 			label: t('PROP_ACTIONS'),
 			options: {
 				filter: false,
-				// display: !props.campaignUnits,
+				display: !noActions,
 				sort: true,
 				customBodyRender: ({ to, id }) => (
 					<React.Fragment>
@@ -100,7 +127,7 @@ function AdUnitsTable(props) {
 								<Visibility color='primary' />
 							</RRIconButton>
 						</Tooltip>
-						{!props.noActions && (
+						{!noClone && (
 							<Tooltip
 								title={t('TOOLTIP_CLONE')}
 								// placement='top'
@@ -129,20 +156,22 @@ function AdUnitsTable(props) {
 	]
 	return (
 		<MUIDataTableEnchanced
-			title={props.campaignUnits ? t('CAMPAIGN_AD_UNITS') : t('ALL_UNITS')}
+			title={campaignUnits ? t('CAMPAIGN_AD_UNITS') : t('ALL_UNITS')}
 			data={data}
 			columns={columns}
 			options={{
 				filterType: 'multiselect',
-				selectableRows: 'none',
-				onRowsSelect: (currentRowsSelected, allRowsSelected) => {
-					console.log(currentRowsSelected, allRowsSelected)
+				rowsSelected: selected,
+				onRowsSelect: (_, allRowsSelected) => {
+					setSelected(allRowsSelected.map(row => row.dataIndex))
 				},
 			}}
 			{...props}
 		/>
 	)
 }
+
+//TEMP
 const truncateString = (string, maxLength = 50) => {
 	if (!string) return null
 	if (string.length <= maxLength) return string
