+++ src/components/dashboard/containers/Tables/tableHelpers.js
@@ -1,0 +1,39 @@
+import React from 'react'
+import DoneAllIcon from '@material-ui/icons/DoneAll'
+import WarningIcon from '@material-ui/icons/Warning'
+import HourglassFullIcon from '@material-ui/icons/HourglassFull'
+import MonetizationOnIcon from '@material-ui/icons/MonetizationOn'
+
+const mapStatusIcons = (humanFriendlyStatus, status, size) => {
+	const icon = {
+		xs: { fontSize: 10 },
+		md: { fontSize: 15 },
+		ls: { fontSize: 20 },
+	}
+	const waitIcon = <HourglassFullIcon style={icon[size]} color={'secondary'} />
+	const doneIcon = <DoneAllIcon style={icon[size]} color={'primary'} />
+	const warningIcon = <WarningIcon style={icon[size]} color={'error'} />
+	const cashIcon = <MonetizationOnIcon style={icon[size]} color={'accentTwo'} />
+	if (humanFriendlyStatus === 'Closed' && status.toLowerCase() !== 'exhausted')
+		return waitIcon
+	switch (status.toLowerCase()) {
+		case 'active':
+		case 'ready':
+			return doneIcon
+		case 'pending':
+		case 'initializing':
+		case 'waiting':
+			return waitIcon
+		case 'offline':
+		case 'disconnected':
+		case 'unhealthy':
+		case 'invalid':
+			return warningIcon
+		case 'withdraw':
+			return cashIcon
+		default:
+			return ''
+	}
+}
+
+export { mapStatusIcons }
